requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
mysqlclient
alembic
python-dotenv
cryptography